flask/requirements.txt
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
charts==0.4.6
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.1
greenlet==1.1.3.post0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.3
opencc-python-reimplemented==0.1.6
outcome==1.2.0
pandas==1.4.3
protobuf==3.20.1
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
Werkzeug==2.2.2
wsproto==1.1.0
yarl==1.8.1
